Release: Home screen created
I started developing the home screen. The user's plants will be displayed. I created an AddPlant screen that allows you to add a new plant to the database. The application stores the user's flowers in Room database. I've rethought the navigation.

TODO:
- I have not yet solved the navigation problem. The selected screen always loads twice.
- Allow the user to upload their own image.
- Alarm class implementation

I will try to finish it this year, Happy Xmas. :)